requirements.txt
selenium==4.33.0
python-dotenv
pandas
openpyxl
